Split state groups into a separate data store (#6296)
