+++ changelog.d/6245.misc
@@ -1,0 +1,1 @@
+Split out state storage into separate data store.
