+++ synapse/storage/data_stores/__init__.py
@@ -15,6 +15,7 @@
 
 import logging
 
+from synapse.storage.data_stores.state import StateGroupDataStore
 from synapse.storage.database import Database, make_conn
 from synapse.storage.engines import create_engine
 from synapse.storage.prepare_database import prepare_database
@@ -54,6 +55,10 @@ class DataStores(object):
                 if "main" in database_config.data_stores:
                     logger.info("Starting 'main' data store")
                     self.main = main_store_class(database, db_conn, hs)
+
+                if "state" in database_config.data_stores:
+                    logger.info("Starting 'state' data store")
+                    self.state = StateGroupDataStore(database, db_conn, hs)
 
                 db_conn.commit()
