+++ synapse/storage/data_stores/state/schema/delta/23/drop_state_index.sql
